back_3d/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
bidict==0.22.1
blinker==1.7.0
cachelib==0.10.2
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
dnspython==2.5.0
eventlet==0.35.0
Flask==3.0.1
Flask-Cors==4.0.0
Flask-Login==0.6.3
Flask-Session==0.6.0
Flask-SocketIO==5.3.6
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
greenlet==3.0.3
gunicorn==21.2.0
h11==0.14.0
httpcore==1.0.2
httpx==0.25.2
idna==3.6
itsdangerous==2.1.2
Jinja2==3.1.3
MarkupSafe==2.1.3
marshmallow==3.20.2
mistralai==0.0.11
orjson==3.9.12
packaging==23.2
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-dotenv==1.0.0
python-engineio==4.8.2
python-socketio==5.11.0
requests==2.31.0
simple-websocket==1.0.0
sniffio==1.3.0
SQLAlchemy==2.0.25
typing_extensions==4.9.0
urllib3==2.1.0
Werkzeug==3.0.1
wsproto==1.2.0
